Revert "no need to check if the jwt exists, because if it doesnt the user wouldnt be able to reach this code"
This reverts commit 0a88660ef5daf6e0a0e1018f6207bfe160a3495e.

+++ App/screens/SettingsScreen/index.tsx
@@ -175,14 +175,16 @@ class SettingsScreen extends Component<ISettingsScreenProps, IISettingsScreenSta
 	}
 
 	private async performSignOut() {
-		console.log(`AsyncStorage jwtToken is true`);
-		try {
-			// Clears out
-			await AsyncStorage.clear();
-			// TODO: This Navigation call for some reason doesn't yet take us back to the Main Screen after signout
-			this.props.navigation.navigate('MainScreen');
-		} catch (ex) {
-			//
+		if (AsyncStorage.getItem('jwtToken')) {
+			console.log(`AsyncStorage jwtToken is true`);
+			try {
+				// Clears out
+				await AsyncStorage.clear();
+				// TODO: This Navigation call for some reason doesn't yet take us back to the Main Screen after signout
+				this.props.navigation.navigate('MainScreen');
+			} catch (ex) {
+				//
+			}
 		}
 	}
